Switch to Typescript

+++ src/lib/courseChannels.ts
@@ -1,18 +1,19 @@
-import { logger } from "../logger.js";
 import _ from "lodash";
-import { Course } from "../models/course.js";
-import { CourseRolesSetting } from "../models/configuration_models/courseRolesSetting.js";
+import { logger } from "../logger";
+import { Course } from "../models/course";
+import { CourseRolesSetting } from "../models/configuration_models/courseRolesSetting";
+import type { Guild, GuildMemberResolvable } from "discord.js";
 
 //import {inspect} from "util";
 
 /**
- * @param {Client} client The Discord bots client.
- * @param {User} user The Discord user in question.
+ * @param {Guild} guild The Discord guild.
+ * @param {GuildMemberResolvable} user The Discord user in question.
  * @param {Course} course The database entry for the course.
  */
-export async function addUserToCourseChannel(guild, user, course) {
+export async function addUserToCourseChannel(guild: Guild, user: GuildMemberResolvable, course: Course) {
     const targetUser = guild.members.resolve(user);
-    await targetUser.roles.add(course.roleId);
+    await targetUser?.roles.add(course.roleId);
 }
 
 /**
@@ -20,9 +21,9 @@ export async function addUserToCourseChannel(guild, user, course) {
  * @param {User} user The Discord user in question.
  * @param {Course} course The database entry for the course.
  */
-export async function removeUserFromCourseChannel(guild, user, course) {
+export async function removeUserFromCourseChannel(guild: Guild, user: GuildMemberResolvable, course: Course) {
     const targetUser = guild.members.resolve(user);
-    await targetUser.roles.remove(course.roleId);
+    await targetUser?.roles.remove(course.roleId);
 }
 
 /**
@@ -30,7 +31,7 @@ export async function removeUserFromCourseChannel(guild, user, course) {
  * @param {string} name The name of the course and the course role.
  * @param {Guild} guild The guild in which the channel is being created for.
  */
-export async function createCourseChannel(name, guild) {
+export async function createCourseChannel(name: string, guild: Guild) {
 
     const courseRoleSettings = await CourseRolesSetting.findOne({ where: { guildId: guild.id } });
     if(courseRoleSettings === null) {
@@ -39,10 +40,9 @@ export async function createCourseChannel(name, guild) {
 
     const joinMessageChannel = await guild.channels.fetch(courseRoleSettings.roleSelectionChannelId);
     
-    // FIXME: This is throwing an error! Why? I really don't know... isTextBased is in fact a documented function and yet it doesn't want to work.
-    /*if(joinMessageChannel.isTextBased() === false) {
+    if (!joinMessageChannel?.isText()) {
         throw new Error("Could not send message to join message channel as it is not text based.");
-    }*/
+    }
 
     name = _.upperCase(name);
     name = name.replace(/ /g, "");
@@ -61,7 +61,7 @@ export async function createCourseChannel(name, guild) {
         VIEW_CHANNEL: false
     });
 
-    const joinMessage = await joinMessageChannel.send(`**${name}** <@&${role.id}>`);
+    const joinMessage = await joinMessageChannel?.send(`**${name}** <@&${role.id}>`);
     // **CS487** @CS487
     await joinMessage.react('👍');
 
@@ -78,9 +78,10 @@ export async function createCourseChannel(name, guild) {
  * Adds an existing course channel to the database.
  * @param {String} channelId 
  * @param {String} joinMessageId 
- * @param {String} roleId 
+ * @param {String} roleId
+ * @param {String} name
  */
-export async function linkExistingCourseChannel(channelId, joinMessageId, roleId, name) {
+export async function linkExistingCourseChannel(channelId: string, joinMessageId: string, roleId: string, name: string) {
     await Course.create({
         name: name,
         channelId: channelId,
@@ -92,8 +93,9 @@ export async function linkExistingCourseChannel(channelId, joinMessageId, roleId
 /**
  * Deletes the course channel and related roles.
  * @param {String} roleId 
+ * @param {Guild} guild
  */
-export async function deleteCourseChannel(roleId, guild) {
+export async function deleteCourseChannel(roleId: string, guild: Guild) {
     const course = await Course.findOne({
         where: { roleId: roleId }
     });
@@ -113,21 +115,25 @@ export async function deleteCourseChannel(roleId, guild) {
     const joinChannelId = courseRoleSetting.roleSelectionChannelId;
     const joinChannel = await guild.channels.fetch(joinChannelId);
 
+    if (!joinChannel?.isText()) {
+        throw new Error("Could not send message to join channel as channel is not text-based.");
+    }
+
     const messageId = course.messageId;
     const joinMessage = await joinChannel.messages.fetch(messageId);
     await joinMessage.delete();
 
     const channelId = course.channelId;
     const channel = await guild.channels.fetch(channelId);
-    await channel.delete();
+    await channel?.delete();
 
     const role = await guild.roles.fetch(roleId);
-    await role.delete();
+    await role?.delete();
 
     await course.destroy();
 }
 
-export async function unlinkExistingCourseChannel(roleId, guild) {
+export async function unlinkExistingCourseChannel(roleId: string, guild: Guild) {
     const course = await Course.findOne({
         where: { roleId: roleId }
     });
